working on implementing proper preprocessor
Right now, the closest thing to a preprocessor is done while compiling each line and won't be able to support macros. I want to change that
