--- compiler/compiler.h
@@ -1,3 +1,0 @@
-
-int compile_error(const char*, ...);
-
