+++ compiler/src/preprocessor.c
@@ -1,0 +1,25 @@
+#include "preprocessor.h"
+#include <stdlib.h>
+#include <string.h>
+
+#include "computils.h"
+
+int preprocess(FILE* input, FILE* processed_input, FILE* statements, int reflect, int verbose) {
+    char rawline[linesize];
+    /* 
+    two goals: 
+    - break down each line into predefined statements
+        - outline the statements found in the statements file to make later steps faster
+        - this is also needed to safely replace labels, macros, and constants
+    - replace all labels, macros, and constants with the equivalent values, and remove comments
+        - start with macros, as they can include other macros, labels, and constants, and they will also affect the address of labels
+            - loop through and expand each macro found
+                - can loop repeatedly and track located macros until no more macros are found
+                - better option might be to recursively expand a macro until the macro is fully expanded
+            - once no more macros exist, replace all constants with predefined values
+            - replace all labels with correct line addresses
+    
+    Final result should only include assembly instructions that can easily be processed by the primary system
+    Support both mirroring and verbosity setttings
+    */
+}
