+++ compiler/src/preprocessor.h
@@ -1,0 +1,5 @@
+#ifndef FILE
+    #include <stdio.h>
+#endif
+
+int preprocess(FILE* input, FILE* processed_input, FILE* statements, int mirror, int verbose);
