+++ compiler/targets/ng-core/ng-core.h
@@ -1,8 +1,6 @@
-#define opcount 8
-#define opcode_size 3
-#define reg_count 3
-#define instruction_size 16
-#define imm_limit 32768
+#ifndef opcount
+    #include "ng-core-defs.h"
+#endif
 
 const char* operators[opcount] = { 
     "+1",
@@ -68,21 +66,9 @@ const char* regtok[reg_count] = {
     "P"
 };
 
-enum bitval {
-    NONE = -1,
-    CI = 0,
-    PTR_IN = 3,
-    OPCODE = 5,
-    ZERO_X = 8,
-    SWAP_XY = 9,
-    OUT_A = 10,
-    OUT_D = 11,
-    OUT_APTR = 12,
-    BRANCH = 13
-};
-
 const enum bitval outbits[reg_count] = {
     OUT_A,
     OUT_D,
     OUT_APTR
 };
+
